requirement.txt
pillow==10.4.0
pypng==0.20220715.0
PyQRCode==1.2.1
